+++ internal/analysis/response.go
@@ -1,0 +1,5 @@
+package analysis
+
+type Response struct {
+	Text string `json:"text"`
+}
